test: split db load tests
The original TestDbLoadWorks was an integration test, so this splits out
the testing into finer-grained pieces.

I'm treating the database implementation as external to the code, which
is why I don't mind loading and querying the database manually in the
tests. Normally, the database is an implementation detail, but here it's
being treated as a kind of protocol.