refactor: load db with transactions
Word of caution: if you use an in-memory database, you can't start a
transaction and run a global query at the same time. I tried to do this
because the `SqliteDB.Select` function is so useful for doing queries,
but I also wanted to have an active transaction. In pseudocode:

```
tx = db.Begin()

/* !! db.Select() will start new connection */
if !db.Select("select role from roles where role = ?", role) {
    tx.Exec("insert into roles (role) values (?)",  role)
}
```

The queries done during a transaction are not usually complicated, and
they usually only return one row, so I'm ok with leaving it to the user
to handle.